Tidied up Code
