+++ Source/Renderer.h
@@ -2,6 +2,8 @@
 #include <iostream>
 #include <SDL.h>
 
+#include "Number.h"
+
 namespace Renderer {
 
 	bool InitializeRenderer();
@@ -10,4 +12,7 @@ namespace Renderer {
 	SDL_Renderer* GetRenderer();
 	SDL_Window* GetWindow();
 	SDL_Surface* GetScreen();
+
+	void DrawPixel(int x, int y, Uint32 pixel);
+	void RenderNumber(Number num, int size = 10);
 } // namespace Renderer
